+++ src/main/java/com/example/services/ProductService.java
@@ -16,7 +16,7 @@ public class ProductService {
     private final ProductRepository productRepository;
 
     @Transactional(readOnly = true)
-    public List<Product> getAllProducts(String category, Double minPrice, Double maxPrice) {
+    public List<Product> getAllProducts(String category, Float minPrice, Float maxPrice) {
         List<Product> products = productRepository.findAll();
         if (category != null && !category.isEmpty()) {
             products = products.stream()
